code to try to run projections
The code takes a sample from the Sea Around Us reconstruction database and runs the GUM package to get the historical status info. Then attempting to run RunProjections. 
Seems to need the 'FindOpenAccess' code, but also can't seem to get the "Sim_Forward" function to run in SnowProjections. 
Is the StatusQuoPolicy necessary in the RunProjections code?